Move types to src
