+++ sources/shared/src/@types/index.d.ts
@@ -1,2 +1,4 @@
 import 'phaser3-docs/typescript/phaser'
 import './serializable'
+export { Level, LevelBackground, Body, Color } from './level'
+export { Sprites, Sprite } from './sprite'
