+++ sources/shared/src/@types/level.d.ts
@@ -1,20 +1,10 @@
+import { Sprites } from './sprite'
+
 export type Color = string
 
 export type LevelBackground = {
   from: Color,
   to: Color
-}
-
-export type Sprites = { [layer: string]: Sprite[] }
-
-export interface Sprite {
-  id?: string
-  x: number,
-  y: number,
-  width: number,
-  height: number,
-  texture: string,
-  frame?: string
 }
 
 export interface Body {
