+++ sources/shared/src/@types/sprite.d.ts
@@ -1,0 +1,11 @@
+export type Sprites = { [layer: string]: Sprite[] }
+
+export interface Sprite {
+  id?: string
+  x: number,
+  y: number,
+  width: number,
+  height: number,
+  texture: string,
+  frame?: string
+}
